Calibration returns double matrix

+++ aiscm/opencv.cpp
@@ -246,14 +246,14 @@ extern "C" {
 
   SCM opencv_camera_calibration(SCM scm_count, SCM scm_sizes, SCM scm_object_points, SCM scm_image_points, SCM scm_image_size)
   {
-    float *camera = (float *)scm_gc_malloc_pointerless(3 * 3 * sizeof(float), "camera-matrix");
-    float *distortion = (float *)scm_gc_malloc_pointerless(5 * sizeof(float), "distortion");
+    float *camera = (float *)scm_gc_malloc_pointerless(3 * 3 * sizeof(double), "camera-matrix");
+    float *distortion = (float *)scm_gc_malloc_pointerless(5 * sizeof(double), "distortion");
     int count = scm_to_int(scm_count);
     std::vector<std::vector<cv::Point3f>> object_points(to_point3f_vector_vector(scm_count, scm_sizes, scm_object_points));
     std::vector<std::vector<cv::Point2f>> image_points(to_point2f_vector_vector(scm_count, scm_sizes, scm_image_points));
     cv::Size image_size(scm_to_int(scm_car(scm_image_size)), scm_to_int(scm_cadr(scm_image_size)));
-    cv::Mat camera_matrix(3, 3, CV_32FC1, camera);
-    cv::Mat dist_coeffs(5, 1, CV_32FC1, distortion);
+    cv::Mat camera_matrix(3, 3, CV_64FC1, camera);
+    cv::Mat dist_coeffs(5, 1, CV_64FC1, distortion);
     std::vector<cv::Mat> rvecs;
     std::vector<cv::Mat> tvecs;
     try {
